SystemUI: Custom Navbar ring targets
This is based off some Team EOS on CNA (props for the nice starting point)
stuff I've been eyeing, with the AOKP touch of course.

This has options for 1-5 custom targets in the nav ring.

When user chooses 4 or 5 targets the ring layout is raised vertically
and the ring diameter is shrunk a bit to accommodate the targets at the 9 and 3.

I've also got layouts in for Small tabs (grouper) .

I need a HDPI device and a full tab (if we have one) to test these and let
 me know if anything needs fixing.

So since there is no git history for this fiasco let me say thanks to
Kevin Liu and CNA for the running head start on this.

  Patch Set 2: Add the GlowPadView.java and TargetDrawable.java patches from
   Danesh M for the custom lockscreen targets originally posted to CM's gerrit
   here http://review.cyanogenmod.com/#/c/19217/

PS3: fix apps not starting if they have been killed before
PS4: Appereantly camera and possibly other apps set the task id to 0
on launch, this checks for that and will launch from home instead
of history if the id is 0.

Change-Id: Icf99f941c8f903e384147cadf72015709f0b62e8

Conflicts:

	packages/SystemUI/res/values/strings.xml
	packages/SystemUI/src/com/android/systemui/SearchPanelView.java