Frameworks: Lazy dev part 2 (1/2)
this is based on the last lazy dev commit from me with the same
reasons, this time navbar and statusbar white color did no stick because
of not used long var to save.

Best on that....I realized later that I am not allone a lazy dev. Aosp
does count -1 automatic as white lmao

whatever changed default valus to -2 and cleaned 4 lines up

cheers

Change-Id: Ie6cb0c80d6b1aff1fdb4c257a658e73506ade4df
Signed-off-by: Daniel <daxx4182@gmail.com>